fix observer module
